+++ CHANGELOG.md
@@ -30,6 +30,8 @@
     - Renamed `behavior::BehaviorHost` to `behavior::Host`.
     - Renamed `behavior::BehaviorPersistence` to `behavior::Persistence`.
 
+    - Renamed `space::SpaceBuilder` to `space::Builder`.
+
 - `all-is-cubes-gpu` library:
     - `in_wgpu::SurfaceRenderer::new()` requires `wgpu::Adapter` instead of `&wgpu::Adapter`.